https://www.realtor.com/research/data/
https://www.realtor.com/research/data/

Residential data
How to use the data: Files are structured for database use, and contain multiple metrics for different levels of geography down to zip code level. ‘Current month data’ shows figures for the latest month available, while the ‘Historical data’ contains monthly series going back in time. The ‘All Residential’ category includes data on all homes (single family home and condo/townhome) for the given level of geography. See key remarks for important info regarding data quality and usage.

Key remarks: Data in this Realtor.com library is based on the most comprehensive and accurate database of MLS-listed for-sale homes in the industry. We aggregate and analyze data from hundreds of sources and produce hundreds of metrics for multiple markets, and curate figures and trends where possible for reliability and comparability. However, as we continue to evolve our coverage and fine-tune our definitions, some data points may be too volatile or incomparable over time or across markets. This is particularly true for data in a) smaller geographies; b) markets with special or changing definitions of active inventory; c) markets with limited or partial listing; and d) coverage markets with limited or partial sales coverage. Where possible, these cases are annotated for individual metrics (see data dictionaries for more info). Also, every month, we reissue the full historical series, and past data points may change as we improve data breadth and accuracy, and/or re-state the data altogether.